+++ src/commands/Fun/luck.js
@@ -15,28 +15,37 @@ class LuckCommand extends Command {
     async messageRun(message) {
         const randomChance = random.int(0, 100);
 
-        let fagBucks = await message.client.FagBucks.findOne({ where: { userId: message.author.id } });
+        let selfFagBucks = await message.client.FagBucks.findOne({ where: { userId: message.author.id } });
 
-        if (!fagBucks) {
+        if (!selfFagBucks) {
             await message.client.FagBucks.create({
                 userId: message.author.id,
-                amount: 100
+                amount: 100,
+                bank: 0
             });
 
-            fagBucks = await message.client.FagBucks.findOne({ where: { userId: message.author.id } });
+            selfFagBucks = await message.client.FagBucks.findOne({ where: { userId: message.author.id } });
         }
 
-        if (randomChance <= 1) {
-            await fagBucks.update({ amount: fagBucks.amount + 1000 });
-            return message.channel.send('Congratulations! You won **1000 💵 FagBucks**!');
+        if (randomChance === 1) {
+            await selfFagBucks.update({ amount: selfFagBucks.amount + 1000 });
+            return message.channel.send('Congratulations! You won **1,000 💵 FagBucks**!');
         } else if (randomChance <= 10) {
-            await fagBucks.update({ amount: fagBucks.amount + 500 });
+            await selfFagBucks.update({ amount: selfFagBucks.amount + 500 });
             return message.channel.send('Congratulations! You won **500 💵 FagBucks**!');
         } else if (randomChance <= 25) {
-            await fagBucks.update({ amount: fagBucks.amount + 250 });
+            await selfFagBucks.update({ amount: selfFagBucks.amount + 250 });
             return message.channel.send('Congratulations! You won **250 💵 FagBucks**!');
         } else {
-            return message.channel.send('Ah, failed! Better luck next time!');
+            const responses = [
+                'Out of luck! The odds were against you.',
+                'Miss! The universe said no.',
+                '**[🎱]** My sources say no.',
+                'Fail! Better luck next time.',
+                'it no work 💀',
+                'Nope! Maybe it\'ll work next time.'
+            ];
+            return message.channel.send(responses[random.int(0, responses.length - 1)]);
         }
     }
 }